Added support for cloudant local
Added support for cloudant local, currently cloudant local runs
in a vagrant VM until it can be dockerized, for the time being
docker commands are by passed for local, with the correct options
added to connect to local, discarding other options

+++ build.rb
@@ -39,17 +39,50 @@ params["couch"] = "couchdb1.6" unless params["couch"]
 #launch docker
 puts "Starting docker container #{$couch}"
 
-unless system("docker run -p 5984:15984 -d --name 'couchdb' #{params["couch"]}")
-	#we need to stop, we failed to run the docker container, just in case we will delete
-	system("docker rm couchdb")
-	exit 1
+        # COUCH_USERNAME = System.getProperty("test.couch.username");
+        # COUCH_PASSWORD = System.getProperty("test.couch.password");
+        # COUCH_HOST  = System.getProperty("test.couch.host","localhost");
+        # COUCH_PORT = System.getProperty("test.couch.port","5984");
+        # HTTP_PROTOCOL = System.getProperty("test.couch.http","http"); //should either be http or https
 
+if params["couch"] == "cloudantlocal" 
+
+	#remove options that will clash with clocal
+	options_to_remove = Array.new
+
+	params[:d_options].each do |d_option|
+		if d_option.start_with?("-Dtest.couch.username") || d_option.start_with?("-Dtest.couch.password") || d_option.start_with?("-Dtest.couch.host") || d_option.start_with?("-Dtest.couch.port")
+			options_to_remove.push d_option
+		end
+	end
+
+	options_to_remove.each do |option|
+		params[:d_options].delete option
+	end
+
+	#add some -d opts to point tests at local instead
+	params[:d_options].push "-Dtest.couch.username=admin"
+	params[:d_options].push "-Dtest.couch.password=pass"
+	params[:d_options].push "-Dtest.couch.host=127.0.0.1"
+	params[:d_options].push "-Dtest.couch.port=8081" #jenkins runs on 8080 this needs to be 8081
+	params[:d_options].push "-Dtest.couch.ignore.auth.headers=true"
+	params[:d_options].push "-Dtest.couch.ignore.compaction=true"
+else
+	unless system("docker run -p 5984:15984 -d --name 'couchdb' #{params["couch"]}")
+		#we need to stop, we failed to run the docker container, just in case we will delete
+		system("docker rm couchdb")
+		exit 1
+
+	end
 end
 
 puts "Performing build"
 #make gradlew executable
 system("chmod a+x ./gradlew ") 
 
+puts params
+
+#exit
 #handle the differences in the platform
 if params["platform"] == "java"
 	system("./gradlew #{params[:d_options].join(" ")} clean check integrationTest")
@@ -60,10 +93,12 @@ end
 #get the build exit code, will exit with this after tearing down the docker container
 exitcode = $?
 
+unless params["couch"] == "cloudantlocal" 
 puts "Tearing down docker container"
 
 system("docker stop couchdb")
 
 system("docker rm couchdb")
+end
 
 exit exitcode.to_i
